Convert uninitialized attributes to Nones and fix newly-introduced index_data bug for ProfileIndex
To keep the type of index_data consistently a dict, a previous commit changed its name to temp_index_data before it is zipped up with index_names to make the dict (as it is usually used). Some instances of self.index_data were unchanged to temp_index_data, which crashed  the read_logs method.