Add magic log/linear data returns.
Now when accessing a linear data category that doesn't exist, but its
logarithmic version does, automatically exponentiate the logarithm
version and return it. Similar for accessing non-existent logarithmic
quantities that have linear counterparts. In the future, may want to
switch to using a safe log for converting linear to logarithmic.