Add in ability to read in mesa models and clean up code
Now `read_data`, which is called by `__init__` will detect if the file is a model file
(.mod) or if it is a standard data file (.log or .data). It will then punt to one of the
two reader methods that will populate the usual data fields from the file.

This is necessary since both files are structured very differently, and right now an
exception is thrown if the extension is none of these types.